// restitution-app/ClientApp/src/model/createVictimEntityRestitutionCaseRequestDto.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Restitution API
 * API for the Restitution Application
 * OpenAPI spec version: v1
 */
import type { CourtInfoDto } from './courtInfoDto';
import type { DocumentDto } from './documentDto';
import type { ParticipantDto } from './participantDto';
import type { VictimEntityApplicationDto } from './victimEntityApplicationDto';

export interface CreateVictimEntityRestitutionCaseRequestDto {
  /** @nullable */
  courtInfoCollection?: CourtInfoDto[] | null;
  /** @nullable */
  providerCollection?: ParticipantDto[] | null;
  /** @nullable */
  documentCollection?: DocumentDto[] | null;
  application?: VictimEntityApplicationDto;
}